Split modal state into multiple files
