+++ Sources/SwiftNavigationCoordinator/Navigator/Modal/ModalState.swift
@@ -14,7 +14,7 @@ final class ModalState {
   fileprivate(set) var _destination: ModalDestinationPath<AnyIdentifiableDestination>?
   
   @PerceptionIgnored
-  private(set) var delegates: [ObjectIdentifier: AnyModalStateDelegate] = [:]
+  fileprivate(set) var delegates: [ObjectIdentifier: AnyModalStateDelegate] = [:]
 
   init<Destination: Sendable & Hashable & Identifiable>(
     initialDestination: ModalDestinationPath<Destination>?
@@ -118,57 +118,3 @@ extension ModalState {
     }
   }
 }
-
-@MainActor
-protocol ModalStateDelegate: AnyObject {
-  func modalStateDidDismiss(_ destination: AnyIdentifiableDestination)
-}
-
-extension ModalStateDelegate {
-  @_disfavoredOverload
-  func eraseToAnyModalStateDelegate() -> AnyModalStateDelegate {
-    AnyModalStateDelegate(self)
-  }
-  
-  func eraseToAnyNavigationQueue() -> AnyModalStateDelegate where Self == AnyModalStateDelegate {
-    self
-  }
-}
-
-@MainActor
-final class AnyModalStateDelegate: ModalStateDelegate {
-  private var _modalStateDidDismiss: ((AnyIdentifiableDestination) -> Void)!
-  
-  private(set) var isValid = true
-  
-  init<Delegate: ModalStateDelegate>(
-    _ delegate: Delegate
-  ) {
-    assert(Delegate.self != AnyModalStateDelegate.self)
-    
-    _modalStateDidDismiss = { [weak self, weak delegate] in
-      guard let delegate else {
-        self?.isValid = false
-        return
-      }
-      
-      delegate.modalStateDidDismiss($0)
-    }
-  }
-  
-  func modalStateDidDismiss(_ destination: AnyIdentifiableDestination) {
-    _modalStateDidDismiss(destination)
-  }
-}
-
-#if canImport(XCTest)
-
-extension ModalState {
-  func testBinding<Destination: Sendable & Hashable & Identifiable>(
-    for destinationType: Destination.Type = Destination.self
-  ) -> Binding<ModalDestinationPath<Destination>?> {
-    Perception.Bindable(self).destination(for: destinationType)
-  }
-}
-
-#endif
